backend/requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
pydantic
tiktoken
google-generativeai
